use opencl version in func tests

+++ src/pruning.h
@@ -1,3 +1,3 @@
-int wont_update_effect(XMatrixSparse X, double lambda, int k, double last_max,
-                       double *last_rowsum, double *rowsum, int *column_cache,
-                       double *beta);
+int wont_update_effect(XMatrixSparse X, float lambda, int k, float last_max,
+                       float *last_rowsum, float *rowsum, int *column_cache,
+                       float *beta);
